requirements.txt
discord.py==2.3.2


PyNaCl


psycopg2-binary

